+++ install.ps1
@@ -39,7 +39,7 @@
     ScriptName : install.ps1
     Purpose    : Bootstrap install / update for AppPackager
     Owner      : CM Engineering
-    Version    : 1.5.2.1
+    Version    : 1.5.2.2
 #>
 
 [CmdletBinding()]